Properly close database connections
Seems we never actually closed the db connections. At all. Oops.

Each meeting keeps a connection, but it should be closed when the
meeting ends. The bigger problem was that we leaked a connection every
time we gave the error that a room was closed, which would quickly eat
all available connections.